requirements.txt
streamlit>=1.32.0
pymongo==4.7.0
bcrypt==4.0.1
python-dotenv==1.0.0
Pillow>=10.0.0
google-generativeai>=0.8.0
fpdf==1.7.2
pandas>=2.1.0
PyMuPDF==1.22.5
